First pass at improving malloc checking in CRAM.
Fixes #920, at least in as far as not causing crashes.

The next step is to go through all the bits of memory we may leak
during the tidyup (although the next thing that'll be happening is
almost certainly cascading up to main and exiting anyway, it's poor
form and it's probably only a 99.9% certainty).