Check kstring memory errors in SAM parser.
Note I modified kputc to return the input character as an unsigned char
(as is done with getc and co).  Without this, "char *s; kputc(*s);" may
return -1 and be interpreted as EOF simply because the input was signed.

Removes unused B tag code (commented out with #if 0 since 829f57).