Report format errors as EFTYPE when available (rather than ENOEXEC)
The BSDs and macOS define EFTYPE ("Inappropriate file type or format"),
so use it to report invalid input file formats instead of the less ideal
but more portable ENOEXEC ("Exec format error").

(tabix.c also checks for ENOEXEC so as to print a better message than
"Exec format error", but strerror(EFTYPE) is already a good message
so there is no need to check for EFTYPE there as well.)

Samtools and bcftools contain no checks against ENOEXEC, so no changes
are needed in those repositories.