Allow library user more control over bam1_t memory ownership
Adds API functions bam_set_mempolicy() and bam_get_mempolicy()
which can be used to set/get the ownership of the bam1_t struct
and the bam1_t::data buffer used for variable-length parts of
the alignment record.

Bitwise flag BAM_USER_OWNS_STRUCT indicates that the bam1_t
structure is being managed outwith HTSlib (for example if it is
an array member).  The only effect is to prevent bam_destroy1()
from trying to free it.

Flag BAM_USER_OWNS_DATA indicates that the user is in control
of the bam1_t::data buffer.  This prevents bam_destroy1() from
trying to free it.  It also changes behaviour when the buffer
is expanded - instead of calling realloc() on the pointer,
a new memory area will be allocated and the existing data
copied into it.  bam1_t::data is updated to point to the new
buffer and the BAM_USER_OWNS_DATA flag is cleared (so
HTSlib takes back management).

BAM_USER_OWNS_DATA can be used to pack variable-length data for
many records into a single buffer.  An example of this is included
in the documentation.

sam_parse1() is changed to use b->data directly instead of copying
the pointer into a kstring and using that instead.  As sam_parse1
was mainly copying bytes directly into the kstring buffer, the
new version isn't much different to the old one.

All bam data resizing is now done with static inlines
realloc_bam_data() and possibly_expand_bam_data().  The
definitions for these have been moved to sam_internal.h so that
they can be used outside of sam.c.

Adds extra tests in test/sam.c for the new functionality.