Add HTS_VERSION #define for use in compile-time #if checks
Add HTS_VERSION macro to aid writing preprocessor checks to test
for the presence of HTSlib facilities, e.g.,

  #if HTS_VERSION >= 101000  // Check for v1.10 or later

This XYYYZZ encoding of X.Y.Z into a single integer is the same encoding
as used by boost/version.hpp and similar. It would be nice to be able to
hide the encoding behind another macro and write tests like

  #if HTS_VERSION >= HTS_VERNUM(1,10,0)  // Doesn't work

but this is a syntax error when compiled against earlier HTSlibs that
don't define HTS_VERNUM().

Like the version numbers in version.sh, NEWS, and the manual pages,
this must be updated manually (or via mkrelease) as part of the commit
that merges develop to master during the release process.

When merging the tagged release commit back to develop, this must be
updated again (manually or via tag_release) as described in the comment
in <htslib/hts.h>.

Fixes #794. (This has been added as 1.9.90, where .90 indicates an
in-development Git repository version leading up to the next release.)