Add bcf_get_info_int64; make bcf_update_info_int64 a static inline
As bcf_get_info_int64() and bcf_update_info_int64() are new
interfaces, they can be made static inlines instead of macros so
that the compiler can check the data type of the values / dst
parameter.  The old macros probably have to stay as they
are as we don't know how they are being used in third-party code.

The documentation around the bcf_get_info_ and bcf_update_info_
macros is made a bit more doxygen-like.