Add support for `@SQ-AN` alternative sequence names
This adds the alternative names into ref_hash without identifying them
as non-canonical names. This means that all queries based on ref_hash
will accept either SN or AN names, even those that explicitly ask for
SN -- this is considered a feature.

As per the existing SN handling in sam_hrecs_update_hashes(), duplicates
produce warnings but are otherwise ignored and the existing ref_hash entry
is left as is. When removing an AN tag, ref_hash entries are only removed
if they do indeed point to the `@SQ` item being altered (otherwise they
are left as is).

[TODO] sam_hdr_update_line() does not yet handle AN altnames.

(View the sam_hrecs_remove_hash_entry() diff with -w as the bulk
of the code is unchanged other than its indentation level.)