Bug fix CRAM 2.1 output.
In tested I spotted v2.1 output still uses the rANS codec, which is
incorrect.  It does this because rans is enabled when opening a file
if the version is >= 3.0, but we set the version after opening the
file and that didn't explicitly disable rans again.

Io_lib's Scramble does these bits in reverse (set default vers and
then open file) so worked fine.