Tweak handling of duplicated @SQ SN: and AN: entries
Spot cases where an @SQ SN: name is the same as an altname
from an earlier @SQ AN: tag, and ensure that the reference
hash table points to the entry using it for SN.  Due to
the way they are read, these clashes would be resolved
correctly if they occurred in SAM or BAM files, but it
was possible to accidentally override the SN name when
reading a CRAM file (making it impossible to select the
masked reference when creating iterators).

Don't complain about duplicates in sam_hrecs_add_ref_altnames()
when the name refers to the correct entry.  This can happen
when SN: is duplicated in AN:, and possibly also if
sam_hrecs_add_ref_altnames gets called more than once for
the same line.

Don't remove the hash table entry in sam_hrecs_remove_ref_altnames
if the name is the primary one (SN:) for the reference.

The "Seen already?" checks in sam_hrecs_update_hashes() are
reordered as the if statements were becoming rather deeply
nested.