Fixed CRAM_OPT_REFERENCE handling for files with reordered headers.
With samtools view -t ref.fa, fp->fn_aux is set.
With samtools view -o reference=ref.fa, this didn't happen.

Both of them load the reference and generate a CRAM_OPT_REFERENCE
parameter, but when writing the CRAM header if fn_aux is set it'll
cause a reload to validate the order in the fai file matches the order
in the header, and if not reorders the structures.

The result of this is using -o to set reference could sometimes get
M5 values out of sync with SQ SNs.