Hide unwanted symbols in Windows dll and add exports check
Change linker option from --export-all-symbols to
--exclude-all-symbols so that exports are restricted to
the symbols that we want.  Only seems to make a difference
for the hts.dll.a implib; the main dll still appears to have
everything.

Define HTS_BUILDING_LIBRARY in the .c files that go
into HTSlib and the ones that make the external plug-ins
so that __declspec(dllexport) annotation only gets added for
the exports.  Apparently __declspec(dllimport) is
not needed as we use the --enable-auto-import linker option.

Add rule to extract exported symbols for the test-shlib-exports
target on MSYS2 / MinGW.

Add detection of MINGW in $(PLATFORM) variable so that using
'make' without 'configure' builds the right sort of dll
when running in an MSYS2 MinGW environment.