Return lengths from bam_cigar2rqlens as hts_pos_t and make it static
Strictly hts_pos_t would only be needed for rlen (and only in very
odd circumstances) but it's easier for both as we would otherwise
need to report a possible overflow in qlen.

bam_cigar2rqlens is not mentioned in any headers and only used
in sam.c, so make it static for now.

Credit to OSS_Fuzz
Fixes oss-fuzz id 18144