Removed some CRAM pre-1.0 legacies.
The preservation map for 1.0 defines MI, UI and PI keys, which in turn
set a bunch of variables related to how quality scores are stored.
The 1.0 spec required this, but even in the official 1.0 cramtools-jar
completely ignored them.

They appear in this C code as I added them while writing io_lib's
implementation from the 1.0 specification, but the variables are never
accessed as that was the only way to get parity with cramtools.jar.

This patch keeps the keys so a 1.0 file doesn't whinge (cramtools did
write those fields out, even if it never did anything with them), but
they are now silently skipped.

[TODO:  ditch 1.0 entirely!  I'm not convinced it works in any
meaningful way, given the specification and Java implementation
differed greatly. CRAM 2.0 was born from this - unifying code and
spec - and IMO should be considered the first workable version.]