Fix call of memcpy with NULL source in sam_dispatcher_read
Fix complaint about a NULL pointer being passed to memcpy,
which happened the first time through the main loop in
sam_dispatcher_read() because kstring 'line' did not have
an allocated buffer at that point.  While harmless as
the length to copy was zero, it caused complaints when
using undefined behaviour sanitizer.

The fix is to call ks_resize() on line at the start of the function
to ensure a buffer has been allocated.  The memory will be needed
later in all likelihood, so this just moves the allocation forward.