Changed int64_t to hts_pos_t for *_parse_region API
Also fixed a few other assumptions of hts_pos_t equalling int64_t.

Updated the test-parse-reg to use hts_pos_t as well.

Added a comment to hts_parse_reg64 indicating that it is not our
preferred API.  It was added in a separate commit for ease of API
migration from hts_parse_reg, but was added only because the
hts_parse_region code hadn't been accepted.  This is now the correct
interface.