Fix htslib.mk parallel build (PR #969)
On Windows and Mac, the shared and static HTSlib libraries depend
in the same .o files.  This dependency was not present in htslib.mk
so a build system using it for a parallel make would try to
build all of the .o files twice if it required both the shared and
static libraries.  It also occasionally resulted in broken
libraries being built due to one of the .o files being updated
while one of the archives was being created.  This behaviour
was spotted for Appveyor Windows builds of bcftools.

Fix by adding an HTSlib Makefile rule that depends on $(LIBHTS_OBJS)
and creates an empty file.  htslib.mk is updated to make this file
a dependency of the static and (where needed) shared libraries
so that the object files get built once before the libraries.
While creating an empty file isn't ideal (for example it won't
notice if a .o file has been removed), it seems to be the best
solution given that htslib.mk can't easily work out the
dependencies between the libraries and the object files (which
can change due to configure options).

Adds htslib.mk dependencies on the static library for the bgzip,
htsfile and tabix targets to prevent a similar problem when
building them in parallel.

Updates the list of files in HTSLIB_ALL.