Silence gcc warning in bcf_hdr_printf()
For some optimisation levels, gcc would report a format-truncation
warning for the line that measured how much memory was needed to
store the header string.  Silence it by supplying a reasonably
big buffer, which has the added bonus that the malloc and second
call to vsnprintf can be avoided completely in most cases.