Fixed CRAM decode on ultra-long cigar.
We can now have many cigar ops (eg ref-skips) that combine to more than 512Mb.

This isn't a full fix for long references as CRAM sequence positions
are still 32-bit.  However this is a starting point down that road and
there are still sizes between 28-bit and 32-bit where the CIGAR failed
to operate correctly.