Add missing Makefile dependencies and ignored test executable
Update dependencies (and reinstate the previous #include "hts_defs.h"
in a couple of headers) so that #includes and dependencies are
consistent with each other. In particular, add $(htslib_hts_defs_h)
where necessary due to PR #946; also in that PR <htslib/sam.h> was
added to textutils.c (so that the declaration of stringify_argv()
is visible to its definition) rather than textutils_internal.h.

PR #930 altered test/sam.c's dependencies.

Add test executable added in PR #952 to .gitignore.