Make 64-bit optional and don't compile by default.
This commit addresses several problems introduced by the recently added
support of 64-bit coordinates:
- 64-bit positions were silently modified on BCF output
- any erroneous out-of-range INFO value resulted in unparseable BCF
  output (in contrast to the previous behavior of a silent under/overflow)
- conversion between VCF and BCF is no longer possible

Large genomes is a niche area and can be easily worked around by splitting
the largest chromosome into two. This commit makes the compilation of 64-bit
values optional, favoring the interroperability of BCF with VCF, a feature
heavily relied on in many pipelines.

The support of large genomes in this commit is limited to POS and INFO tags
defined as Number=1 (motivated by INFO/END), all other extreme values
are replaced with a missing value ("."). This changes the original ad-hoc
approach which supported first field of any INFO tag array.

The default 32-bit mode adds checks to prevent under/overflow of INFO and
FORMAT integer values and prints a warning when extreme values are replaced
with a missing value. An additional check is added to prevent overflow of
64-bit coordinates on BCF output.

Note that the 64-bit support is not compiled in by default because:
- it breaks VCF/BCF specification
- it works only for VCF but cannot produce a functional BCF
- there is currently no front-end capable of indexing VCFs with 64bit
  coordinates (n_lvls is not accessible via `bcftools index`)

Note that a full 64-bit support for BCF would require significantly more
work and large changes of the codebase, API, and the BCF specification,
therefore it is not planned at the moment.

The limited POS and INFO/END support for VCF can be compiled in by
passing the ALLOW_INT64 macro to the compiler.

Make the warning messages more informative.
Additional checks to prevent writing invalid BCF containing 64-bit data.
Temporarily comment out failing tests which assume 64 bit positions
are compiled in by default.