Use including-file-relative paths to public htslib/*.h headers
The full preprocessor search path for #include "..." is typically:

    1) directory of the file containing the #include directive
    2) environment variables $CPATH, $C_INCLUDE_PATH, etc
    3) command line options -I etc
    4) system include directories

When a top-level *.[ch] file #includes "htslib/foo.h", the nearby
header file within the source tree is found due to (1). When a file
in a subdirectory (htslib/*.h, cram/*.[ch], or test/*.c) does the
same #include, the nearby header file is intended to be found due
to (3) via the hard-coded `-I.` in the Makefile. (`.` here being the
compiler process's $PWD, i.e., the top-level HTSlib source directory.)

However the latter can be subverted if there is an "htslib" directory
containing foo.h in $CPATH, $C_INCLUDE_PATH, or on the command line
prior to the hard-coded `-I.`. This will be the case if users have
made a previous HTSlib installation accessible via $CPATH/etc or by
adding a -I option to $(CFLAGS) (rather than $(CPPFLAGS)).

Avoid this by adding ../ to #includes in subdirectories (and always
using "" rather than <>), so that all includes of htslib/*.h within
HTSlib are found due to (1). Fixes #347.