Use TYPEKEY() for all header type code arithmetic
Change the existing TYPEKEY() macro to a static inline so that the
input and output types are more accurately recorded. Take care about
casting the plain chars to unsigned, to avoid sign extension bugs and
compiler and cppcheck warnings. (This new code results in the same or
better assembly code as the previous tests and macro.)

Also rename a sam_hdr_remove_lines() parameter [trivial], as the
declaration in htslib/sam.h uses different parameter prefixes and
the coincidence of different h parameters caused a cppcheck warning.