Fixed nuances between gzip vs bgzf compressed files.
Credit to OSS-Fuzz
Fixes oss-fuzz 20473

This also fixes the indexing tests forbidding samtools index on
gzipped or naked SAM and BAMs.  Previously the index command did not
fail, but gave broken indices.

Ironically with 1.10 samtools index on a totally raw uncompressed BAM
did actually start working (but not gzipped BAM or raw/gzipped SAM).
This was not by design and the indices produced couldn't be used by
1.9 and prior so it's safest to forbid this case too.

Also added documentation on the ambiguous "is_bgzf" field given this
is set for BAM files, even if ungzipped.  It's purely a flag on usage
of the bgzf_open/read functions, which can read uncompressed data, and
has nothing to do with the file format actually being BGZF.