Make the iterator work with 64 bit positions. (PR #1018)
Remove the usage of hts_pair64_max_t::max, as the same value
is already contained in hts_reglist_t::max_end.
Remove references to hts_pair64_max_t type.

Note that while this changes the data type of hts_itr_t::off,
and so strictly changes the API/ABI, the array pointed to
should only be accessed by the iterator functions in HTSlib
itself and not by external code.  As the change will be
invisible to code using the iterator interfaces, the library
SOVERSION has not been changed.