Added bam_set_seqi function.
This has been copied over from the samtools sort_minimiser PR.
For now this is a macro to keep the code style consistent with all the
other bam_* macros.

Ideally we'd use a static inline, but if so I think this is a second
PR (or second commit) that rewrites all these macro-functions so we
have some level of consistency in code style.