Add bcf to vcf test
This is a better test for the reader than vcf->bcf->vcf, which
may incorrectly work due to matching bugs in the reader and
writer.  This is especially likely on big-endian platforms.

test/tabix/vcf_file.vcf is reused for the purpose.  It gets
a ##FILTER=<ID=PASS..> line so the reader doesn't have to add one.