Fix integer overflow in cram_read_slice
Occurs when block content_id is INT_MIN, which cannot be
negated in an int.

Casting to uint32_t fixes the problem and allows checks for
negative values to be eliminated.  Blocks with negative
content_id will end up in different hash buckets, but that
shouldn't matter as long as cram_get_block_by_id() is
consistent with cram_read_slice().

Credit to OSS-Fuzz
Fixes oss-fuzz 20102