Fixes threaded bgzf --write-index.
This adds the analogue of the hts_idx_amend_last function for bgzf.
This is necessary when multi-threading output using --write-index.

Fixes samtools/samtools#1197

In theory the change should have no impact as the only difference is
whether our virtual offset points to the end of a block or the start
of the next block.  Either way the two offsets are essentially the
same locaiton on disk.  However due to a bug elsewhere (see next
commit) this lead to unreported bgzf_read failures.