Permit bgzf block offsets to be at the end of a block.
This fixes a bug caused by indices generated with
"samtools view --write-index -@8" (fixed in previous commit).  This
was leaving some block offsets at the end of the current block instead
of the start of the next block.

However bgzf_read treated such scenarios as a truncated read of length
0, which the calling code then interpreted as EOF.

Now at the exact end of a block is valid and triggers reading the next
block, but beyond the end of a block is treated as an erroneous block
offset.