Try to handle BCF files with negative rlen fields.
This can happen where VCF files with invalid END tag values (less
than the reference position) have been converted to BCF.

bcf_read1_core() is changed to treat rlen as a signed value (which
is what the BCF2 specification says).  This means the maximum
rlen now supported by HTSlib for BCF files is reduced to 2^31.

If rlen is negative, bcf_record_check() will now print out a
warning.  It will also attempt to fix up rlen by substituting
the length of the reference allele.

A call to bcf_record_check() is added to bcf_readrec() so
reads using iterators will be protected from bad input data.
Unfortunately checks involving the header have to be disabled
for this interface as it isn't available.