Fix last commit by reporting #ref_id instead of ref name.
It turns out with various multi-threading options it can occasionally
happen that the reference name has been freed already (or not yet
allocated).  Travis spotted it, but on 1000s of runs I've been unable
to reproduce this locally.  It's only diagnostics though so safest bet
is simply to report numerics instead.