Make bgzf_read_block() print more error messages
To make it easier to find out where BGZF reads have failed.
It prints the offset of the BGZF block where the read / decompress
failed, which makes it much easier to locate bad parts of files.
Unfortunately it can't report the file name as it's not stored
by bgzf or hfile, but hopefully one of the callers will divulge
it.

This introduces a small change of behaviour as block_address is
now updated after an empty block has been read.  This may change
the location stored in indexes of files with empty blocks,
however the index should be more efficient as it will now point
to the location where data starts instead of the previous
empty block.