Add tabix --cache option; make bgzf cache work better with hfile_libcurl (#1053)
* Add tabix bgzf block caching.

* Preserve buffer contents for deferred seek in hfile_libcurl.

Commit ad2e24f added delayed seeks to hfile_libcurl so that
calling hseek() lots of times without an hread() in between would
not cause unnecessary web requests (the bgzf block cache tends to
do this).  While it mostly worked, a small amount of data that
has been read but not consumed yet was being dropped by hseek().
Where this data was actually needed after all the hseek()ing,
it would cause hfile_libcurl() to have to rewind slightly resulting
in a new web request.  Again, this can frequently happen when
using the bgzf block cache.

To prevent this, make hfile_libcurl take a copy of the data that
would have been lost.  If a later read with a deferred seek hits
the preserved region, it can be returned from the cached data
without having to start a new web request.  Once this cached
data has been used up, the original web connection can carry
on reading from where it left off.

* Fix libcurl_read() when not all preserved data is used up

fp->preserved_bytes should not be altered.

Assertion at line 780 needs to be changed as it's no longer
true that fp->base.begin and fp->base.end will always point to
the start of the buffer.