Use including-file-relative paths to top-level private headers
Similarly to the previous commits, includes of these headers from
files in subdirectories can be subverted if $CPATH/etc or $(CFLAGS)
activates a directory containing them. This is unlikely,  but the
particularly misguided might perhaps add a previous HTSlib source
directory to $C_INCLUDE_PATH.

To avoid that scenario, and for consistency with all the other
intra-HTSlib #includes, add ../ to #includes of top-level headers
in subdirectories.

With this, the only remaining #includes in HTSlib requiring `-I.` (as
is hard-coded in Makefile) are each *.c file's <config.h> inclusion.
(Using <> for that is as recommended by the autoconf documentation.)