Update to the depth filtering for mpileup.
We can now count the depth anywhere along the read rather than only at
the left end.  It's not buffering reads up so it doesn't know what's
about to come, so for any individual column it'll still favour reads
that end near a column than start near a column, but by counting at
the right hand end we avoid the boom & bust approach to before where
the coverage could plumet to zero immediately following a large spike.

With iter->depth_pos = 1.0 the coverage ends up being a minimum of
maxcnt (assuming the true coverage is higher), rather than a maximum
of maxcnt.  So we go over-depth, but avoid the drop outs.