Improve bam_aux_update_str()
Makes bam_aux_update_str() easier to use by not insisting that
the byte count passed in via the len parameter should include
the NUL-terminator for the string.  This makes it less likely
that an invalid tag will be generated by forgetting to include
the NUL, and also allows tags to be made from part of a longer
string.  To support this, it is made to handle adding new tags
itself instead of delegating to bam_aux_append().

Where existing tags are replaced, bam_aux_update_str() is
changed to more accurately calculate how much extra space is
required.  A call to bam_aux_del() is removed and the tag
shuffling code updated to reduce the number of memmove() calls
needed to get the following tags in the right place.

Adds some extra tests to ensure growing and shrinking tags
both work correctly.  Also fixes a bug in the test/sam.c
test_mempolicy() function, which was not adding the correct
number of bytes when appending the ZZ tag to each alignment
record.