Better catch BCF CHROM/FILTER/FORMAT ids absent from the header
It is possible to make a sparse set of ids in the VCF
header using IDX= annotations.  This can cause problems
if a BCF record refers to one of the indexes that does
not exist.

Update bcf_record_check() so that these bad records while
files are being read.  It already did this for INFO; this
adds similar checks for CHROM, FILTER and FORMAT data.

Credit to OSS-Fuzz
Fixes oss-fuzz 23670