Permit format string ubam and ucram for uncompressed data.
Eg samtools view -O ubam in.sam | ...

Getting this to handle file suffixes would be tricky.  Currently
sam_open_mode parses filename extensions to turn .bam, .cram, .sam or
.sam.gz into mode string b, c, "" or z, but supporting .ubam would
need "b0" which means the caller has to handle longer mode strings.
That's not true without changing other code, and we've never explained
how long the mode should be as it was always liimted to 1 char in the
past.