Catch invalid dictionary indexes in vcf_format()
bcf_record_check() can't check the validity of references to the
header dictionaries when using the iterator interface as it doesn't
have access to the header.  Add a check to vcf_format() to catch any
invalid records that make it that far.

Changes the existing, incomplete,  check for FORMAT ids so
that it returns -1 instead of calling abort().

Adds 'const' qualifier on 'rec' to bcf_seqname() and
bcf_seqname_safe() they can be used without casts in vcf_format().