Avoid crash for `bcf_update_alleles(hdr, rec, NULL, 0)`
Updating a bcf1_t to have 0 alleles is not particularly meaningful, but
it is a valid state as that's what bcf_init() and bcf_clear() produce.
Hence it is a valid thing for bcf_update_alleles() to do too, so calling
it with nals==0 should not crash. Instead set rlen the same way that
bcf_init() and bcf_clear() leave it. Fixes #994.

Also fix bcf_get_format_values() allocation failure check [minor].
Hat tip @reedacartwright.