Rename public HTSlib struct tags to omit leading underscores
File-scope identifiers starting with an underscore are reserved to
the compiler, and there is nothing wrong with a struct tag matching
a typedef name.

In cases where foo_t is actually a typedef to a pointer, rename the
struct tag from __foo_t to foo_s to distinguish it from foo_t.