Fixed a (harmless) CRAM uninitialised data access.
This was caused by the fix in PR #1094.

The problem there was data without explicit quality scores being set
(CF data series with "preserve qual scores" not set) could be coupled
with explicit Q or q feature codes to add qualities at specific
sites.  Without quals, the default is "*" (a run of Q255), but we
can't just amend a few qualities here and there as that leads to a mix
of Q255 and Q-something-else.  We memset from Q255 to Q-something-else
the on the first explicit quality change.

The flaw is that this check is applied even on data where the
"preserve qual scores" CF value is set, but we haven't initialised the
quality string to all-Q255 in that scenario.

However it was totally harmless, as if we preserve quality scores,
then the entire array of quality values are then subsequently
overwritten (after applying any per-base qualities - dumb I know, but
it was never intended for that scenario).  So irrespective of the
outcome of the uninitialised check and whether or not the memset to
Q30 happened, the data still ends up the same thing.

Note curiously this does not show up with -fsanitize=address, but
valgrind does detect it.  Hence the test harness was passing fine.