Fix bug in the line reallocation for threaded SAM decoding.
(The +NM/2 on the wrong side of the comparison!)

The +8 changes here may not be necessary, but we did it for some
allocs so I'm just being consistent and making sure it's always got 8
more than it uses.  Either it's not necessary in which case those
changes are harmless, or it is necessary in which case we were
sometimes not always leaving at least 8 bytes remaining.

I believe this fixes https://github.com/samtools/samtools/issues/1293