Don't trash errno in bam_tag2cigar.
bam_aux_get sets errno=ENOENT when we look for a tag and don't find
it.  Restore it to the old value when failing to find CG tags as
not finding the tag is not an error.

The consequence of samtools/samtools#1296 was that some errors
found while decoding a BAM file got reported as "No such file
or directory" prior to this commit.

Co-Authored-By: Rob Davies <rmd+git@sanger.ac.uk>