Improved return from bcf_hdr_parse_line to enable error detection.
The function returns NULL for both failure and also for running out of
text.  Hence it's executed in a loop until it returns NULL, assuming
success.

On successful end it also sets *len to 0, so now on error we
explicitly set *len to non-zero so it is possible to use that as
a side channel for disambiguating EOT from error.  Values returned
in *len are documented in the header file.

If bcf_hdr_parse_line() returns NULL due to a malformed line,
bcf_hdr_parse() can use the returned value in *len to skip to
the next one instead of searching for the newline itself.

To stop multiple warning messages from being printed on finding
malformed lines, make hts_hdr_parse_line() responsible for logging
both cases that it traps and remove the corresponding logging from
hts_hdr_parse().  Use hts_strprint() when printing bad lines to
sanitise any unprintable characters.

Co-Authored-By: Rob Davies <rmd+git@sanger.ac.uk>