Improve clean-up in hts_tpool_init()
Allocate hts_tpool::t_stack earlier so it's easier to clean up
if it fails.

Add error handling code to tidy up any threads that got created
prior to a pthread_create() failure.  Rename variable i to
t_idx so it's easier to see how it links between the thread creation
and the clean-up code.  Adds a dependency on htslib/hts_log.h so
it can use hts_log_error().