Bug fix to the pileup constructor.
We call constructor and destructor for each new bam object that is
added to and removed from the pileup.  Pileup also takes its own copy
of the bam1_t struct as sam_read1 loops may be using the same
structure over and over again.

However the constructor was called with the passed in bam1_t instead
of pileups own copy of it (unlike destructor).  Hence anything the
constructor attempts to cache, such as the location of an aux tag,
may get written over.