Improvement (personal preference) to previous C99 fix.
In #498 I removed the warning:

    vcf.c:3030:26: warning: assignment makes pointer from integer without a cast [-Wint-conversion]
                 d->allele[i] = tmp.l;

by storing a pointer starting from a fixed "" string, and later on
subtracting it back again to get the length.  While that works, it's
pretty icky and I think undefined behaviour as it uses pointer
arithmetic outsides the bounds of an object.  (In reality it's fine on
all modern systems as the days of segmented pointer archictures are
long gone.)

Instead I now cast the length into a char* as the size of allele is
never going to grow beyond the size of a pointer, and then cast back
prior to adding to the d->als pointer.

[Ideally we wouldn't need these shenanigans at all and keep it in index
space instead of pointers so growing memory doesn't require a lot of
pointer juggling, but bcf_dec_t is a public structure.]