Add sam_format_aux1 API.
This is extracted out from sam_format1_append1, and indeed is the
lion's share of this function.  It is useful as an external API for
anything that is generating its own tags or a partial tag list in SAM
format, such as "samtools fastq".

Although not used as such, the tag key to use and the data (as
returned from bam_aux_get) are separate pointers.  This permits
renaming of tags, but also avoids the oddities of why bam_aux_get
returns 2 bytes into the in-memory representation of a tag.